Added unfinished e-mail sending thing for the coin

+++ src/main/java/com/springboot/jobseeker/shared/data/model/ProfileDetail.java
@@ -3,6 +3,8 @@ package com.springboot.jobseeker.shared.data.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name="profile_details")
 @Builder
@@ -19,7 +21,7 @@ public class ProfileDetail extends MasterEntity {
     private String address;
     private String summary;
     private String cvUrl;
-    private Integer totalCoin;
+    private BigDecimal totalCoin;
     private String companyName;
 
     @ManyToOne(fetch = FetchType.LAZY)
